slightly faster, unsigned ints

+++ Cython DTW/PythonDTWTiming.py
@@ -19,8 +19,8 @@ c2 = np.random.normal(0,1,(2000,2000))
 
 # setup parameters
 
-dn = np.array([1,1,2,1], np.int16)
-dm = np.array([1,2,1,3], np.int16)
+dn = np.array([1,1,2,1], np.uint32)
+dm = np.array([1,2,1,3], np.uint32)
 dw = np.array([1.0, 1.0, 2.0, 3.0])
 subsequence = True
 parameter = {'dn': dn, 'dm': dm, 'dw': dw, 'SubSequence': subsequence}
@@ -53,8 +53,8 @@ print("Total: %f"%(np.mean(secondStepEndTime - startTime)))
 
 # setup parameters
 
-dn = np.array([1,1,2,1], dtype=np.int16)
-dm = np.array([1,2,1,3], dtype=np.int16)
+dn = np.array([1,1,2,1], dtype=np.uint32)
+dm = np.array([1,2,1,3], dtype=np.uint32)
 dw = np.array([1.0, 1.0, 2.0, 3.0])
 subsequence = False
 parameter = {'dn': dn, 'dm': dm, 'dw': dw, 'SubSequence': subsequence}
